+++ src/rendering.ts
@@ -1,5 +1,12 @@
-import {Vector2, WebGLRenderer} from "three";
+import {OrthographicCamera, Scene, Vector2, WebGLRenderer} from "three";
 import { EffectComposer } from "three/examples/jsm/postprocessing/EffectComposer.js";
+import { RenderPixelatedPass } from "three/examples/jsm/postprocessing/RenderPixelatedPass.js";
+import {isClient} from "./util";
+import {invariant} from "./Error";
+import {VIEWPORT_SIZE} from "./constants";
+
+declare const canvas: HTMLCanvasElement;
+
 /**
 * @class NullComposer
 * A drop in replacement for the Three.js EffectComposer that does nothing.
@@ -27,3 +34,41 @@ export class NullRenderer extends WebGLRenderer {
   domElement = null as unknown as HTMLCanvasElement;
 }
 
+export function createRenderer() {
+  if(!isClient) {
+    return new NullRenderer();
+  }
+  invariant(
+    canvas instanceof HTMLCanvasElement,
+    `Missing canvas element with id "canvas"`
+  );
+  const renderer = new WebGLRenderer({
+    canvas,
+    antialias: false,
+    precision: "lowp",
+    powerPreference: "low-power"
+  });
+  renderer.setSize(VIEWPORT_SIZE.x, VIEWPORT_SIZE.y);
+  // We want these to be set with CSS
+  Object.assign(canvas.style, {
+    width: "",
+    height: ""
+  });
+
+  return renderer;
+}
+
+export function createEffectComposer(
+  renderer: WebGLRenderer,
+  scene: Scene,
+  camera: OrthographicCamera
+) {
+  const composer = new EffectComposer(renderer);
+  const pixelatedPass = new RenderPixelatedPass(2, scene, camera, {
+    depthEdgeStrength: -0.5,
+    normalEdgeStrength: -1
+  });
+  composer.addPass(pixelatedPass);
+
+  return composer;
+}
