+++ src/state/index.ts
@@ -3,7 +3,7 @@ import { Texture, Scene, AnimationMixer, Vector2 } from "../Three";
 import { World } from "../EntityManager";
 import { IEntityPrefab } from "../EntityPrefab";
 import { EntityPrefabEnum, IEntityPrefabState } from "../entities";
-import { createEffectComposer, createOrthographicCamera, createRenderer } from "../systems/RenderSystem";
+import { createOrthographicCamera } from "../systems/RenderSystem";
 import { menuRoute } from "../routes";
 import { Observable, ObservableArray, ObservableSet } from "../Observable";
 import { Behavior } from "../systems/BehaviorSystem";
@@ -25,7 +25,7 @@ import { SystemManager } from "../System";
 import { BehaviorEnum } from "../behaviors";
 import { LoadingItem } from "../systems/LoadingSystem";
 import {IEditorState} from "../systems/EditorSystem";
-import {NullComposer} from "../rendering";
+import {createEffectComposer, createRenderer, NullComposer} from "../rendering";
 
 // Create Object abstraction inspired by Pharo & Koi. Focus on
 // - Composability: compose complex objects out of basic objects. Basic objects represent a single value/type and give it a name. Use valueOf or toString to convert them to primatives.
